Complete test for 'left.applier' function factory
